general cleanup, better API, error improvements, and more
