+++ app_distribution_server/routers/health_router.py
@@ -1,0 +1,12 @@
+from fastapi import APIRouter
+from fastapi.responses import PlainTextResponse
+
+router = APIRouter(tags=["Healthz"])
+
+
+@router.get(
+    "/healthz",
+    response_class=PlainTextResponse,
+)
+async def healthz() -> PlainTextResponse:
+    return PlainTextResponse(content="OK")
