+++ templates/error.jinja.html
@@ -2,6 +2,6 @@
 
 {% block content %}
     <main>
-        <h1>404 - Not found</h1>
+        <h1>{{ error_message }}</h1>
     </main>
 {% endblock %}
